# lib/api-zod/src/generated/types/certificateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerPilot career management API
 * OpenAPI spec version: 0.1.0
 */

export interface CertificateInput {
  name: string;
  organization: string;
  date: string;
  credentialUrl: string;
}