# lib/api-zod/src/generated/types/certificateUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerPilot career management API
 * OpenAPI spec version: 0.1.0
 */
import type { CertificateInput } from './certificateInput';

export type CertificateUpdate = CertificateInput;