# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerPilot career management API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  title: string;
  description: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  targetDate: string;
}