# lib/api-zod/src/generated/types/goalUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CareerPilot career management API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalInput } from './goalInput';

export type GoalUpdate = GoalInput;